Consolidate info data access to HomeScene, add video posts
